Reduce code complexity in GoogleChart.
