+++ assets/js/components/GoogleChart/utils.js
@@ -1,0 +1,197 @@
+/**
+ * Google Chart utilities.
+ *
+ * Site Kit by Google, Copyright 2023 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * External dependencies
+ */
+import set from 'lodash/set';
+import cloneDeep from 'lodash/cloneDeep';
+import merge from 'lodash/merge';
+
+/**
+ * Internal dependencies
+ */
+import { stringToDate } from '../../util/date-range/string-to-date';
+
+/**
+ * Returns the modified data for Google chart.
+ *
+ * @since n.e.x.t
+ *
+ * @param {Array|null}      data          Chart data.
+ * @param {Array|undefined} selectedStats If only certain columns should be displayed for the data set.
+ * @return {Object} The modified data to use in a chart.
+ */
+export const getGoogleChartData = ( data, selectedStats ) => {
+	// Ensure we don't filter out columns that aren't data, but are things like
+	// tooltips or other content.
+	let nonDataColumns = [];
+	if ( data?.length ) {
+		nonDataColumns = data[ 0 ].reduce( ( acc, row, rowIndex ) => {
+			return row?.role ? [ ...acc, rowIndex ] : acc;
+		}, [] );
+	}
+
+	let modifiedData = data;
+	if ( selectedStats?.length > 0 ) {
+		modifiedData = data.map( ( row ) => {
+			return row.filter( ( _columnValue, columnIndex ) => {
+				return (
+					columnIndex === 0 ||
+					selectedStats.includes( columnIndex - 1 ) ||
+					nonDataColumns.includes( columnIndex - 1 )
+				);
+			} );
+		} );
+	}
+
+	return modifiedData;
+};
+
+/**
+ * Returns the optimal height and width for a preview element.
+ *
+ * @since n.e.x.t
+ *
+ * @param {string|undefined} loadingHeight Preview height.
+ * @param {string|undefined} height        Chart height.
+ * @param {string|undefined} loadingWidth  Preview width.
+ * @param {string|undefined} width         Chart width.
+ * @return {Object} The optimal height and width to use in a preview element.
+ */
+export const getLoadingDimentions = (
+	loadingHeight,
+	height,
+	loadingWidth,
+	width
+) => {
+	const dimentions = {
+		height: loadingHeight || height,
+		width: loadingWidth || width,
+	};
+	// If a loading height is set but a width is not (or a loading width is set
+	// but not a height), change the "unset" value to 100% to avoid visual bugs.
+	// See: https://github.com/google/site-kit-wp/pull/2916#discussion_r623866269
+	if ( dimentions.height && ! dimentions.width ) {
+		dimentions.width = '100%';
+	}
+	if ( dimentions.height && ! dimentions.width ) {
+		dimentions.height = '100%';
+	}
+
+	return dimentions;
+};
+
+/**
+ * Returns a combined chart events object.
+ *
+ * @since n.e.x.t
+ *
+ * @param {Array.<Object>|null} chartEvents Event names and its callbacks.
+ * @param {Function|undefined}  onReady     Chart event.
+ * @param {Function|undefined}  onSelect    Chart event.
+ * @return {Object} The object containig all events.
+ */
+export const getCombinedChartEvents = ( chartEvents, onReady, onSelect ) => {
+	const combinedChartEvents = [ ...( chartEvents || [] ) ];
+
+	if ( onReady ) {
+		combinedChartEvents.push( {
+			eventName: 'ready',
+			callback: onReady,
+		} );
+	}
+
+	if ( onSelect ) {
+		combinedChartEvents.push( {
+			eventName: 'select',
+			callback: onSelect,
+		} );
+	}
+
+	return combinedChartEvents;
+};
+
+/**
+ * Returns a chart configuration object.
+ *
+ * @since n.e.x.t
+ *
+ * @param {Object}  options       Configuration data.
+ * @param {boolean} gatheringData If chart is in gathering info state.
+ * @param {string}  chartType     Chart types: PieChart, LineChart.
+ * @param {string}  startDate     Start date for a user data range.
+ * @param {string}  endDate       End date for a user data range.
+ * @return {Object} Chart options configuration.
+ */
+export const getChartOptions = (
+	options,
+	gatheringData,
+	chartType,
+	startDate,
+	endDate
+) => {
+	const chartOptions = cloneDeep( options );
+	if ( gatheringData && chartType === 'LineChart' ) {
+		if ( ! options?.vAxis?.viewWindow?.min ) {
+			set( chartOptions, 'vAxis.viewWindow.min', 0 );
+		}
+		if ( ! options?.vAxis?.viewWindow?.max ) {
+			set( chartOptions, 'vAxis.viewWindow.max', 100 );
+		}
+		if ( ! options?.hAxis?.viewWindow?.min ) {
+			set(
+				chartOptions,
+				'hAxis.viewWindow.min',
+				stringToDate( startDate )
+			);
+			delete chartOptions.hAxis.ticks;
+		}
+		if ( ! options?.hAxis?.viewWindow?.max ) {
+			set(
+				chartOptions,
+				'hAxis.viewWindow.max',
+				stringToDate( endDate )
+			);
+			delete chartOptions.hAxis.ticks;
+		}
+	}
+
+	merge( chartOptions, {
+		hAxis: {
+			textStyle: {
+				fontSize: 10,
+				color: '#5f6561',
+			},
+		},
+		vAxis: {
+			textStyle: {
+				color: '#5f6561',
+				fontSize: 10,
+			},
+		},
+		legend: {
+			textStyle: {
+				color: '#131418',
+				fontSize: 12,
+			},
+		},
+	} );
+
+	return chartOptions;
+};
